EnemyAi base
Needed to add AI navigation to the project. The enemy now moves and looks at what has the player tag. The enemy does not have the ability to shoot due to me seeing there is a gun and bullet script and I'm not sure how they work. could not test the take damage. added a basic material for flash when damage is taken. commented Shoot function and if !isShooting.