added moving platforms and ground triggers to platforms
Needed to rewrite how the player determines if they are on the ground. We basically have a script for the ground triggers that set the player's grounded variable on or off when it enters and leaves the trigger. Also had to fix some issues it caused with dashing and other stuff.